booking-opt-prod/optimize-plotter-function/requirements.txt
functions-framework==3.*
google-cloud-storage
matplotlib
numpy
cloudevents